+++ lib/main.dart
@@ -7,7 +7,8 @@ import 'package:txdx/screens/home_screen.dart';
 import 'package:txdx/screens/settings_screen.dart';
 import 'package:window_size/window_size.dart';
 
-import 'providers/shared_preferences_provider.dart';
+import 'providers/settings_provider.dart';
+import 'settings.dart';
 import 'theme/theme.dart';
 
 Future<void> main() async {
@@ -20,8 +21,15 @@ Future<void> main() async {
   }
 
   runApp(
-    const ProviderScope(
-        child: TxDxApp()),
+    ProviderScope(
+      child: Consumer(builder: (context, ref, _) {
+        final settingsFuture = ref.watch(settingsFutureProvider);
+        return settingsFuture.maybeWhen(
+          data: (d) => const TxDxApp(),
+          orElse: () => const CircularProgressIndicator(),
+        );
+      }),
+    ),
   );
 }
 
@@ -32,17 +40,23 @@ class TxDxApp extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final namespace = ref.watch(namespaceProvider);
-    final appTitle = namespace == 'release'  ? 'TxDx' : 'TxDx - Debug';
+    final appTitle = namespace == 'release' ? 'TxDx' : 'TxDx - Debug';
 
     if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
       setWindowTitle(appTitle);
     }
+
+    final useSystemTheme = ref.watch(settingsProvider).getBool(settingsThemeUseSystem);
+    final useDarkTheme = ref.watch(settingsProvider).getBool(settingsThemeUseDark);
+
+    final themeMode = useSystemTheme ? ThemeMode.system : (useDarkTheme ? ThemeMode.dark : ThemeMode.light);
 
     return GetMaterialApp(
       title: appTitle,
       debugShowCheckedModeBanner: namespace != 'release',
       theme: TxDxTheme.light(),
       darkTheme: TxDxTheme.dark(),
+      themeMode: themeMode,
       initialRoute: '/',
       routes: {
         '/': (context) => const HomeScreen(),