+++ lib/screens/settings_screen.dart
@@ -2,11 +2,10 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:txdx/input/browser.dart';
-import 'package:txdx/providers/file_notifier_provider.dart';
-import 'package:txdx/providers/shared_preferences_provider.dart';
 
-import '../providers/file_change_provider.dart';
+import '../settings.dart';
 import '../providers/platform_info_provider.dart';
+import '../providers/settings_provider.dart';
 import '../widgets/menu_header_widget.dart';
 
 class SettingsScreen extends ConsumerWidget {
@@ -30,6 +29,8 @@ class SettingsScreen extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final namespace = ref.watch(namespaceProvider);
     final appVersion = ref.watch(appVersionProvider);
+
+    final usingSystemTheme = ref.watch(settingsProvider).getBool(settingsThemeUseSystem);
 
     return Material(
       child: Column(
@@ -67,20 +68,16 @@ class SettingsScreen extends ConsumerWidget {
                             crossAxisAlignment: CrossAxisAlignment.end,
                             children: [
                               Consumer(builder: (context, ref, _) {
-                                final filenameNotifier = ref.watch(todoTxtFilenameProvider);
-                                return filenameNotifier.map(
-                                  data: (data) => Text(
-                                      (data.value == null) ? 'No file selected' : data.value!
-                                  ),
-                                  loading: (_) => const CircularProgressIndicator(),
-                                  error: (_) => const Text('Error'),
-                                );
+                                final filename = ref.watch(settingsProvider).getString(settingsFileTodoTxt);
+                                return Text(
+                                      filename ?? 'No file selected'
+                                  );
                               }),
                               TextButton(
                                 child: const Text('📂 Select file'),
                                 onPressed: () => {
                                   _pickFile().then((filename) {
-                                    ref.read(todoTxtFilenameProvider.notifier).setFilename(filename ?? '');
+                                    ref.read(settingsProvider).setString(settingsFileTodoTxt, filename ?? '');
                                   })
                                 },
                               ),
@@ -102,20 +99,16 @@ class SettingsScreen extends ConsumerWidget {
                               crossAxisAlignment: CrossAxisAlignment.end,
                               children: [
                                 Consumer(builder: (context, ref, _) {
-                                  final filenameNotifier = ref.watch(archiveTxtFilenameProvider);
-                                  return filenameNotifier.map(
-                                    data: (data) => Text(
-                                        (data.value == null) ? 'No file selected' : data.value!
-                                    ),
-                                    loading: (_) => const CircularProgressIndicator(),
-                                    error: (_) => const Text('Error'),
-                                  );
+                                  final filename = ref.watch(settingsProvider).getString(settingsFileArchiveTxt);
+                                  return Text(
+                                        filename ?? 'No file selected'
+                                    );
                                 }),
                                 TextButton(
                                   child: const Text('📂 Select file'),
                                   onPressed: () => {
                                     _pickFile().then((filename) {
-                                      ref.read(archiveTxtFilenameProvider.notifier).setFilename(filename ?? '');
+                                      ref.read(settingsProvider).setString(settingsFileArchiveTxt, filename ?? '');
                                     })
                                   },
                                 ),
@@ -137,15 +130,61 @@ class SettingsScreen extends ConsumerWidget {
                             crossAxisAlignment: CrossAxisAlignment.end,
                             children: [
                               Switch(
-                                  value: ref.watch(fileAutoReloadNotifierProvider),
+                                  value: ref.watch(settingsProvider).getBool(settingsFileAutoReload),
                                   onChanged: (value) {
-                                    ref.read(fileAutoReloadNotifierProvider.notifier).setAutoReload(value);
+                                    ref.watch(settingsProvider).setBool(settingsFileAutoReload, value);
                                   }
                               ),
                             ]
                           )
                         )
                       ]
+                    ),
+
+                    TableRow(
+                        children: [
+                          const Padding(
+                            padding: EdgeInsets.all(8.0),
+                            child: Text('Use system theme brightness'),
+                          ),
+                          Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.end,
+                                  children: [
+                                    Switch(
+                                        value: ref.watch(settingsProvider).getBool(settingsThemeUseSystem),
+                                        onChanged: (value) {
+                                          ref.read(settingsProvider).setBool(settingsThemeUseSystem, value);
+                                        }
+                                    ),
+                                  ]
+                              )
+                          )
+                        ]
+                    ),
+
+                    TableRow(
+                        children: [
+                          const Padding(
+                            padding: EdgeInsets.all(8.0),
+                            child: Text('Use system dark theme'),
+                          ),
+                          Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.end,
+                                  children: [
+                                    Switch(
+                                        value: ref.watch(settingsProvider).getBool(settingsThemeUseDark),
+                                        onChanged: !usingSystemTheme ? (value) {
+                                          ref.read(settingsProvider).setBool(settingsThemeUseDark, value);
+                                        } : null,
+                                    ),
+                                  ]
+                              )
+                          )
+                        ]
                     ),
                   ]
                 ),