+++ lib/settings.dart
@@ -1,0 +1,13 @@
+const settingsFileTodoTxt = 'todotxt_filename';
+const settingsFileArchiveTxt = 'archivetxt_filename';
+const settingsFileAutoReload = 'file_autoreload';
+const settingsThemeUseSystem = 'theme_system_brightness';
+const settingsThemeUseDark = 'theme_use_dark';
+
+const defaultSettings = {
+  settingsFileTodoTxt: null,
+  settingsFileArchiveTxt: null,
+  settingsFileAutoReload: true,
+  settingsThemeUseSystem: true,
+  settingsThemeUseDark: false,
+};